+++ src/modules/Heightmap/UI/styles.css
@@ -1,0 +1,4 @@
+#canvasOvr {
+    border: 1px solid #000000;
+  }
+
