+++ src/modules/UI/UI.css
@@ -44,10 +44,6 @@
   right: 0px;
 }
 
-#canvasOvr {
-  border: 1px solid #000000;
-}
-
 #assetUrl {
   top: 30px;
 }
@@ -70,22 +66,4 @@
 
 .inputBtn:hover {
   opacity: 0.9;
-}
-
-.inputTextForm {
-  width: 50%;
-  height: 35%;
-  margin: 5%;
-}
-
-#textArea {
-  width: 100%;
-  height: 100%;
-  /* margin: 5%; */
-}
-
-
-.submitBtn {
-  width: 100%;
-  text-align: end;
 }
